backend/retrieval/requirements.txt
cohere>=4.5.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0